:art: restructure and rename component in frontend
