+++ frontend/src/components/home/page/HomePage.tsx
@@ -4,15 +4,15 @@ const ContentWithSidebarLayout = React.lazy(
   () => import('../../commons/layouts/ContentWithSidebarLayout')
 );
 const BannerWrapper = React.lazy(() => import('../components/BannerWrapper'));
-const PromotionWrapper = React.lazy(() => import('../components/PromotionWrapper'));
-const ProductWrapper = React.lazy(() => import('../components/ProductWrapper'));
+const NowPromotionWrapper = React.lazy(() => import('../components/NowPromotionWrapper'));
+const LatestProductWrapper = React.lazy(() => import('../components/LatestProductWrapper'));
 
 const HomePage: FC = () => {
   return (
     <ContentWithSidebarLayout>
       <BannerWrapper />
-      <PromotionWrapper />
-      <ProductWrapper />
+      <NowPromotionWrapper />
+      <LatestProductWrapper />
     </ContentWithSidebarLayout>
   );
 };
