+++ frontend/src/components/promotion/components/PromotionCard.tsx
