+++ frontend/src/graphql/availablePromotionQuery.ts
