+++ frontend/src/graphql/filterProductQuery.ts
