+++ frontend/src/graphql/latestProductQuery.ts
