+++ frontend/src/graphql/nowPromotionQuery.ts
