+++ frontend/src/graphql/productBySlugQuery.ts
